Implementation of the new build support.
Now the static and shared versions are built if possible.
The default installed version can be configured using the
new DEFAULT_LIBRARY_TYPE make variable. The setup is now
persistent using a generated makefile.setup makefile.
The shared version is built only if supported by the
plateform.
--
Work for H505-007.