Fast implementation of the "in" operator.
For the "in" operator it is needed to walk through all values
in the set. To speed-up this an hash set is created with
all values for a composite tag. This is done the first time an
"in" operator is used on a composite tag. There is no time
penality if no "in" operator is used.
--
Continued work for I527-042.