Add a tree simplifier.
This simplifier removes all dead conditional branches. This makes
the tree lighter and faster to interpret. This is also useful to
have a minimal memory footprint. A macro can have many conditionals
to take into account optional arguments. When the macro is inlined
those branches are freed if not used.

Adjust tree output of some tests due to simplifications. Add four
new regression tests.
--
For J802-015.